% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{do_fit_model}
\alias{do_fit_model}
\title{Fit model}
\usage{
do_fit_model(
  tweets,
  stem = get_valid_stems(),
  overwrite = TRUE,
  ...,
  overwrite_tune = overwrite,
  overwrite_fit = overwrite,
  overwrite_preds = overwrite,
  overwrite_shap = overwrite
)
}
\arguments{
\item{tweets}{Tweets from \code{import_tweets()}}

\item{stem}{Either \code{"favorite"} or \code{"retweet"}}

\item{overwrite}{Whether to overwrite existing fit, predictions, and SHAP values.}

\item{...}{Extra arguments to pass to \code{.transform_tweets()}}

\item{overwrite_tune, overwrite_fit, overwrite_pred, overwrite_shap}{Specific booleans for overwriting specific outputs saved to file. Default is to use same value as \code{overwrite}.}
}
\description{
Fit model, make predictions, and generate SHAP values.
}
