% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_get.R
\name{do_get}
\alias{do_get}
\title{Wrapper function for importing and exporting}
\usage{
do_get(
  ...,
  f = NULL,
  file = tempfile(),
  ext = "rds",
  dir = getwd(),
  path = NULL,
  f_import = rio::import,
  f_export = rio::export,
  append = FALSE,
  export = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{...}{Extra arguments to pass to \code{f}}

\item{f}{function to execute}

\item{file}{File name (without extension) to generate \code{path} if \code{path} is not explicitly provided.}

\item{ext}{File extension to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{dir}{Directory to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{path}{Path to export to.}

\item{f_import}{Function to import with if file exists and \code{overwrite = TRUE}.}

\item{f_export}{Function to export with if \code{export = TRUE} .}

\item{append}{Whether to append. Supersedes \code{export} and \code{overwrite}.}

\item{export}{Whether to export. Supersedes `overwrite.}

\item{overwrite}{Whether to overwrite.}
}
\description{
Wrapper that imports and exports result of a function to a file on disk.
}
