% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_favorite}
\alias{fit_favorite}
\title{Fitted \code{{xgboost}} model for count of favorites}
\format{
An object of class \code{xgb.Booster} of length 13.
}
\usage{
fit_favorite
}
\description{
Fitted \code{{xgboost}} model for count of favorites
}
\keyword{datasets}
