% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_retweet}
\alias{fit_retweet}
\title{Fitted \code{{xgboost}} model for count of retweets}
\format{
An object of class \code{xgb.Booster} of length 13.
}
\usage{
fit_retweet
}
\description{
Fitted \code{{xgboost}} model for count of retweets
}
\keyword{datasets}
