% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{get_cols_lst}
\alias{get_cols_lst}
\title{List of columns specifying use in model.}
\usage{
get_cols_lst(stem = get_valid_stems())
}
\arguments{
\item{stem}{Either \code{"favorite"} or \code{"retweet"}}
}
\description{
List of columns specifying use in model.
}
