% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_tweets.R
\name{get_twitter_token}
\alias{get_twitter_token}
\title{Get Twitter token}
\usage{
get_twitter_token(...)
}
\arguments{
\item{...}{Passed to \code{rtweet::create_token()}}
}
\description{
Get Twitter token
}
\details{
Environment variables named "TWITTER_APP|CONSUMER_API_KEY..." are expected to exist.
}
