% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{.import_tweets}
\alias{.import_tweets}
\alias{import_tweets}
\title{Import tweets}
\usage{
.import_tweets(
  user = "xGPhilosophy",
  n = 3200,
  ...,
  dir = .get_dir_data(),
  file = sprintf("\%s_timeline", user),
  ext = "rds",
  path = NULL,
  f_import = readr::read_rds,
  f_export = readr::write_rds,
  append = TRUE,
  export = TRUE,
  overwrite = FALSE
)

import_tweets(train = TRUE, ...)
}
\arguments{
\item{user}{User for whom to retrieve tweets for.}

\item{n}{Number of tweets to retrieve}

\item{...}{Extra parameters passed to \code{rtweet::get_timeline()}}

\item{dir}{Directory to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{file}{File name (without extension) to generate \code{path} if \code{path} is not explicitly provided.}

\item{ext}{File extension to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{path}{Path to export to.}

\item{f_import}{Function to import with if file exists and \code{overwrite = TRUE}.}

\item{f_export}{Function to export with if \code{export = TRUE} .}

\item{append}{Whether to append. Supersedes \code{export} and \code{overwrite}.}

\item{export}{Whether to export. Supersedes `overwrite.}

\item{overwrite}{Whether to overwrite.}

\item{train}{Whether to re-pull all tweets (\code{TRUE}), or just grab the newest (\code{FALSE}).}
}
\description{
Import tweets from xGPhilosophy
}
\details{
You probably don't want to change any of the parameters of this function manually.

This is a wrapper around the non-exported \code{.import_tweets()}. \code{append}, \code{export}, and \code{overwrite} are determined for you based on \code{train}.
}
\section{Functions}{
\itemize{
\item \code{.import_tweets}: Import tweets from xGPhilosophy
}}

