% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict_new}
\alias{predict_new}
\title{Make predictions}
\usage{
predict_new(
  tweets,
  stem = get_valid_stems(),
  overwrite = TRUE,
  ...,
  .overwrite = list(preds = overwrite, shap = overwrite)
)
}
\arguments{
\item{tweets}{Tweets from \code{import_tweets()}}

\item{stem}{Either \code{"favorite"} or \code{"retweet"}}

\item{overwrite}{Whether to overwrite existing fit, predictions, and SHAP values.}

\item{...}{Extra arguments to pass to \code{.transform_tweets()}}

\item{.overwrite}{Specific booleans for overwriting specific outputs saved to file. Default is to use same value as \code{overwrite}.}
}
\description{
Make predictions for new tweets, appending to existing predictions and SHAP values.
}
