% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tweet.R, R/retrieve_tweets.R,
%   R/screenshot_latest_tweet.R
\name{generate_tweet}
\alias{generate_tweet}
\alias{retrieve_tweets}
\alias{screenshot_latest_tweet}
\title{Generate a tweet}
\usage{
generate_tweet(pred, tweets, ..., user = .get_user_bot(), dry_run = TRUE)

retrieve_tweets(
  method = .get_valid_tweet_methods(),
  user = .get_user(),
  ...,
  tweets = NULL,
  n = 3200,
  dir = .get_dir_data(),
  file = sprintf("\%s_timeline", user),
  ext = "rds",
  path = NULL,
  f_import = readr::read_rds,
  f_export = readr::write_rds,
  export = TRUE,
  overwrite = FALSE
)

screenshot_latest_tweet(
  tweets = NULL,
  status_id = NULL,
  user = .get_user(),
  dir = .get_dir_data(),
  file = sprintf("\%s_latest_tweet", user),
  ext = "png",
  path = NULL,
  ...
)
}
\arguments{
\item{pred}{Data frame with \verb{\{stem\}_pred}, \verb{\{stem\}_pred_prnk}, \verb{(tm|g|xg)_(h|a)}, and \code{created_at} columns.}

\item{tweets}{This can be tweets retrieved from something like \code{retrieve_tweets()} or just a data frame with at least \code{created_at} and \code{status_id}.}

\item{...}{Extra parameters passed to \code{tweetrmd::tweet_screenshot()}}

\item{user}{User for whom to retrieve tweets for. (xGPhilophy by default.)}

\item{dry_run}{Whether or not to actually make a tweet.}

\item{method}{How to retrieve tweets.}

\item{n}{Number of tweets to retrieve. (3200 by default.)}

\item{dir}{Directory to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{file}{File name (without extension) to generate \code{path} if \code{path} is not explicitly provided.}

\item{ext}{File extension to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{path}{Path to export to.}

\item{f_import}{Function to import with if file exists and \code{overwrite = TRUE}.}

\item{f_export}{Function to export with if \code{export = TRUE} .}

\item{export}{Whether to export. Supersedes `overwrite.}

\item{overwrite}{Whether to overwrite.}
}
\description{
Generate a tweet

Import tweets from xGPhilosophy

Screenshot xGPhilosophy's latest tweet
}
\section{Valid \code{method}s}{

\itemize{
\item \code{"none"} return existing tweets at \code{path}
\item \code{"all"} re-trieve and return entire timeline
\item \code{"since"} retrieve timeline since last tweet saved at \code{path} and return it along with existing tweets (functionally this will return the same results as \code{"all"} , unless a tweet has since been deleted)
\item \code{"new"} only retrieve and return timeline since last tweet saved at \code{path}
}
}

