% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_tweets.R
\name{transform_tweets}
\alias{transform_tweets}
\title{Transform tweets}
\usage{
transform_tweets(tweets, ..., train = TRUE, first_followers_count = 5000)
}
\arguments{
\item{...}{Not currently used}

\item{train}{If \code{TRUE}, then updates team follower account numbers. Otherwise, uses an extrapolated based on the last retrieved numbers.}

\item{tweet}{Tweets retrieved with \code{retrieve_tweets}}

\item{first_followers_account}{Assumed number of xGPhilospher followers around the end of 2019. We have to make an assumption so that we can do interpolation of growth of followers.}
}
\description{
Transform tweets into format for modeling. This can be saved and combined with SHAP values later. Follower growth of the xGPhilophy account is assumed to be linear per tweet, while growth of the team accounts is assumed to follow a concave curve.
}
