% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_predict.R
\name{do_predict}
\alias{do_predict}
\title{Make predictions}
\usage{
do_predict(
  tweets_transformed,
  stem = get_valid_stems(),
  overwrite = TRUE,
  ...,
  .overwrite = list(preds = NULL, shap = NULL)
)
}
\arguments{
\item{tweets_transformed}{Tweets retrieved from \code{retrieve_tweets()} and passed into \code{transform_tweeets()}}

\item{stem}{Either \code{"favorite"} or \code{"retweet"}}

\item{overwrite}{Whether to overwrite existing fit, predictions, and SHAP values.}

\item{...}{Unused}

\item{.overwrite}{Specific booleans for overwriting specific outputs saved to file. Default is to use same value as \code{overwrite}. Only use this if you know what you're doing.}
}
\description{
Make predictions for new tweets, appending to existing predictions and SHAP values.
}
