% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweets-generate.R
\name{generate_tweet}
\alias{generate_tweet}
\title{Generate a tweet}
\usage{
generate_tweet(
  pred,
  tweets,
  in_reply_to_tweets,
  preds,
  preds_long,
  shap_long,
  ...,
  prelude = NULL,
  appendix = NULL,
  user = .get_user_bot(),
  dry_run = TRUE,
  delete_plot = !dry_run
)
}
\arguments{
\item{pred}{Data frame with \verb{\{stem\}_pred}, \verb{\{stem\}_pred_prnk}, \verb{(team|g|xg)_(h|a)}, and \code{created_at} columns.}

\item{tweets}{Tweets of the user (punditratio by default) who will be making a reply. Needed to check that a tweet hasn't already been made.}

\item{in_reply_to_tweets}{Tweets of the user for whom a tweets will be made in reply to (xGPhilosophy by default). Needed for identifying the tweet to reply to.}

\item{preds}{Predictions with at least \code{team_h}, \code{team_a}, \verb{\{stem\}_count}.}

\item{preds_long}{Tidy predictions with at least five columns \code{status_id}, \code{stem} (either favorite or retweet), \code{pred}, and \code{count}. This will be used to generate a plot to accompany the tweet.}

\item{shap_long}{Tidy SHAP values with at least four columns \code{status_id}, \code{stem} (either favorite or retweet), \code{lab}, and \code{shap_value}. This will be used to generate a plot to accompany the tweet.}

\item{...}{Extra parameters passed to .\code{save_plot()}.}

\item{prelude, appendix}{Extra text to add to beginning or end of auto-generated tweet.}

\item{user}{Who is making the reply. (pundit_ratio by default.)}

\item{dry_run}{Whether or not to actually make a tweet.}
}
\description{
Generate a tweet
}
