% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_tweets.R
\name{.retrieve_tweets}
\alias{.retrieve_tweets}
\alias{retrieve_tweets}
\title{Import tweets}
\usage{
.retrieve_tweets(
  user = "xGPhilosophy",
  n = 3200,
  ...,
  dir = .get_dir_data(),
  file = sprintf("\%s_timeline", user),
  ext = "rds",
  path = NULL,
  f_import = readr::read_rds,
  f_export = readr::write_rds,
  append = TRUE,
  export = TRUE,
  overwrite = FALSE
)

retrieve_tweets(method = c("none", "all", "since", "new"), ...)
}
\arguments{
\item{user}{User for whom to retrieve tweets for.}

\item{n}{Number of tweets to retrieve}

\item{...}{Extra parameters passed to \code{rtweet::get_timeline()}}

\item{dir}{Directory to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{file}{File name (without extension) to generate \code{path} if \code{path} is not explicitly provided.}

\item{ext}{File extension to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{path}{Path to export to.}

\item{f_import}{Function to import with if file exists and \code{overwrite = TRUE}.}

\item{f_export}{Function to export with if \code{export = TRUE} .}

\item{append}{Whether to append. Supersedes \code{export} and \code{overwrite}.}

\item{export}{Whether to export. Supersedes `overwrite.}

\item{overwrite}{Whether to overwrite.}

\item{method}{How to retrieve tweets.}
}
\description{
Import tweets from xGPhilosophy
}
\details{
This is a wrapper around the non-exported \code{.retrieve_tweets()}. \code{append}, \code{export}, and \code{overwrite} are determined for you based on \code{method}.
}
\section{Valid methods}{

\itemize{
\item \code{"none"} return existing tweets at \code{path}
\item \code{"all"} re-trieve and return entire timeline
\item \code{"since"} retrieve timeline since last tweet saved at \code{path} and return it along with existing tweets (functionally this will return the same results as \code{"all"} , unless a tweet has since been deleted)
\item \code{"new"} only retrieve and return timeline since last tweet saved at \code{path}
}
}

