% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_tweets.R
\name{retrieve_tweets}
\alias{retrieve_tweets}
\alias{retrieve_tweets_auto}
\title{Import tweets}
\usage{
retrieve_tweets(
  method = .get_valid_tweet_methods(),
  ...,
  tweets = NULL,
  user = "xGPhilosophy",
  n = 3200,
  dir = .get_dir_data(),
  file = sprintf("\%s_timeline", user),
  ext = "rds",
  path = NULL,
  f_import = readr::read_rds,
  f_export = readr::write_rds,
  append = TRUE,
  export = TRUE,
  overwrite = FALSE
)

retrieve_tweets_auto(method = .get_valid_tweet_methods(), ...)
}
\arguments{
\item{method}{How to retrieve tweets.}

\item{...}{Extra parameters passed to \code{rtweet::get_timeline()}}

\item{tweets}{If not \code{NULL}, then this assumed to have been retrieved by the user beforehand (e.g. by calling \code{retrieve_tweets(export = FALSEE)}). New tweets will not be scraped. Rather, they will simply be appended to existing tweets (assuming \code{append = TRUE}).}

\item{user}{User for whom to retrieve tweets for. (xGPhilophy by default.)}

\item{n}{Number of tweets to retrieve. (3200 by default.)}

\item{dir}{Directory to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{file}{File name (without extension) to generate \code{path} if \code{path} is not explicitly provided.}

\item{ext}{File extension to use to generate \code{path} if \code{path} is not explicitly provided.}

\item{path}{Path to export to.}

\item{f_import}{Function to import with if file exists and \code{overwrite = TRUE}.}

\item{f_export}{Function to export with if \code{export = TRUE} .}

\item{append}{Whether to append. Supersedes \code{export} and \code{overwrite}.}

\item{export}{Whether to export. Supersedes `overwrite.}

\item{overwrite}{Whether to overwrite.}
}
\description{
Import tweets from xGPhilosophy
}
\details{
This is a wrapper around the non-exported \code{retrieve_tweets()}. \code{append}, \code{export}, and \code{overwrite} are determined for you based on \code{method}.
}
\section{Valid \code{method}s}{

\itemize{
\item \code{"none"} return existing tweets at \code{path}
\item \code{"all"} re-trieve and return entire timeline
\item \code{"since"} retrieve timeline since last tweet saved at \code{path} and return it along with existing tweets (functionally this will return the same results as \code{"all"} , unless a tweet has since been deleted)
\item \code{"new"} only retrieve and return timeline since last tweet saved at \code{path}
}
}

